LPS-79131 Fix typo
