+++ modules/apps/collaboration/message-boards/message-boards-demo-data-creator-impl/src/main/java/com/liferay/message/boards/demo/data/creator/internal/LiferayMBCategoryDemoDataCreatorImpl.java
@@ -14,8 +14,7 @@
 
 package com.liferay.message.boards.demo.data.creator.internal;
 
-import com.liferay.message.boards.constants.MBCategoryConstants;
-import com.liferay.message.boards.demo.data.creator.RootMBCategoryDemoDataCreator;
+import com.liferay.message.boards.demo.data.creator.MBCategoryDemoDataCreator;
 import com.liferay.message.boards.model.MBCategory;
 import com.liferay.petra.string.CharPool;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -33,14 +32,14 @@ import org.osgi.service.component.annotations.Component;
  * @author Sergio González
  */
 @Component(
-	property = {"source=liferay"}, service = RootMBCategoryDemoDataCreator.class
+	property = {"source=liferay"}, service = MBCategoryDemoDataCreator.class
 )
-public class LiferayRootMBCategoryDemoDataCreatorImplImpl
+public class LiferayMBCategoryDemoDataCreatorImpl
 	extends BaseMBCategoryDemoDataCreatorImpl
-	implements RootMBCategoryDemoDataCreator {
+	implements MBCategoryDemoDataCreator {
 
 	@Override
-	public MBCategory create(long userId, long groupId)
+	public MBCategory create(long userId, long parentCategoryId)
 		throws IOException, PortalException {
 
 		int randomElement = RandomUtil.nextInt(
@@ -49,16 +48,18 @@ public class LiferayRootMBCategoryDemoDataCreatorImplImpl
 		String name = _names.get(randomElement);
 		String description = _descriptions.get(randomElement);
 
+		MBCategory category = mbCategoryLocalService.getMBCategory(
+			parentCategoryId);
+
 		return createCategory(
-			userId, groupId, MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID,
-			name, description);
+			userId, category.getGroupId(), parentCategoryId, name, description);
 	}
 
 	private static List<String> _read(String fileName) {
 		return Arrays.asList(
 			StringUtil.split(
 				StringUtil.read(
-					LiferayRootMBCategoryDemoDataCreatorImplImpl.class,
+					LiferayMBCategoryDemoDataCreatorImpl.class,
 					"dependencies/liferay/" + fileName + ".txt"),
 				CharPool.NEW_LINE));
 	}
